requirements.txt
google-cloud-dialogflow==2.25.0
google-generativeai==0.2.1
streamlit==1.28.0